buddy_backend/requirements.txt

cat > buddy_backend/requirements.txt <<'EOF'
fastapi
uvicorn
python-dotenv
google-generativeai
EOF
